requirements.txt
Django==5.0.3
openpyxl==3.1.2
pandas==2.2.1
numpy==1.26.4
python-docx==1.1.0